Reactive 기능 지원을 위해 sample 모듈 세분화
