+++ documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/emitter/RestAssuredDocumentEmitter.kt
@@ -59,7 +59,7 @@ class RestAssuredDocumentEmitter(
                 response.statusCode,
                 response.fields.associatedFieldSample()
             )
-            val mockMvc = controller(api).build(provider)
+            val mockMvc = controller(api).buildMockMvc(provider)
             val requestSpecification: MockMvcRequestSpecification = given().mockMvc(mockMvc)
 
             requestSpecification
