+++ documentify-sample/build.gradle.kts
@@ -9,20 +9,32 @@ jacoco {
 
 tasks {
     jacocoTestReport {
-        dependsOn("test")
+        dependsOn(
+            "${rootProject.projects.documentifySample.mvcSample.identityPath.path}:test",
+            "${rootProject.projects.documentifySample.reactiveSample.identityPath.path}:test"
+        )
 
         val coreModulePath = rootProject.projects.documentifyCore.identityPath.path
         val coreModuleClasses = project(coreModulePath).sourceSets.getByName("main").output.classesDirs
         val coreModuleSources = project(coreModulePath).sourceSets.getByName("main").allSource
+
         additionalClassDirs.setFrom(files(coreModuleClasses))
         sourceDirectories.setFrom(files(coreModuleSources))
+
+        executionData.setFrom(
+            project(":documentify-sample:mvc-sample").layout.buildDirectory.file("jacoco/test.exec"),
+            project(":documentify-sample:reactive-sample").layout.buildDirectory.file("jacoco/test.exec")
+        )
 
         reports {
             html.required.set(true)
             xml.required.set(true)
             csv.required.set(false)
         }
-        finalizedBy("jacocoTestCoverageVerification")
+        finalizedBy(
+            "${rootProject.projects.documentifySample.mvcSample.identityPath.path}:jacocoTestCoverageVerification",
+            "${rootProject.projects.documentifySample.reactiveSample.identityPath.path}:jacocoTestCoverageVerification"
+        )
         classDirectories.setFrom(files(classDirectories.files.map {
             fileTree(it) {
                 exclude("**/documentify/sample/**")
@@ -40,13 +52,10 @@ tasks {
     }
 }
 
-dependencies {
-    implementation(libs.spring.boot.starter.web)
-    implementation(libs.spring.boot.starter.test)
-    testImplementation(libs.mockk)
-    testImplementation(projects.documentifyCore)
-}
-
-tasks.test {
-    useJUnitPlatform()
+subprojects {
+    dependencies {
+        implementation(rootProject.libs.spring.boot.starter.test)
+        testImplementation(rootProject.libs.mockk)
+        testImplementation(rootProject.projects.documentifyCore)
+    }
 }
