+++ documentify-sample/mvc-sample/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    jacoco
+}
+
+jacoco {
+    toolVersion = "0.8.12"
+}
+
+dependencies {
+    implementation(rootProject.libs.spring.boot.starter.web)
+}
