+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/UnnestedSampleController.kt
@@ -1,0 +1,19 @@
+package io.github.bgmsound.documentify.sample.mvc
+
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/unnested-sample")
+class UnnestedSampleController {
+    @GetMapping
+    fun unnestedSample(): String {
+        return "unnested sample"
+    }
+
+    @GetMapping("/list")
+    fun unnestedSampleList(): List<String> {
+        return listOf("unnested sample 1", "unnested sample 2")
+    }
+}
