+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/UseCaseBindingController.kt
@@ -1,0 +1,25 @@
+package io.github.bgmsound.documentify.sample.mvc
+
+import io.github.bgmsound.documentify.sample.mvc.dto.response.SampleResponse
+import org.springframework.stereotype.Service
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/use-case")
+class UseCaseBindingController(
+    private val useCase: UseCase
+) {
+    @GetMapping
+    fun useCaseBinding(): SampleResponse {
+        return useCase.doSomethingAndReturn()
+    }
+}
+
+@Service
+class UseCase {
+    fun doSomethingAndReturn(): SampleResponse {
+        return SampleResponse(1, "usecase binding")
+    }
+}
