+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/dto/request/NestedSampleRequest.kt
@@ -1,0 +1,6 @@
+package io.github.bgmsound.documentify.sample.mvc.dto.request
+
+data class NestedSampleRequest(
+    val nestedField: SampleRequest,
+    val listedField: List<SampleRequest>
+)
