+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/dto/request/SampleRequest.kt
@@ -1,0 +1,6 @@
+package io.github.bgmsound.documentify.sample.mvc.dto.request
+
+data class SampleRequest(
+    val integerField: Int,
+    val stringField: String
+)
