+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/dto/response/ComplexSampleResponse.kt
@@ -1,0 +1,5 @@
+package io.github.bgmsound.documentify.sample.mvc.dto.response
+
+data class ComplexSampleResponse(
+    val nestedField: NestedSampleResponse
+)
