+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/dto/response/NestedSampleResponse.kt
@@ -1,0 +1,6 @@
+package io.github.bgmsound.documentify.sample.mvc.dto.response
+
+data class NestedSampleResponse(
+    val nestedField: SampleResponse,
+    val listedField: List<SampleResponse>
+)
