+++ documentify-sample/mvc-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/mvc/dto/response/SampleResponse.kt
@@ -1,0 +1,6 @@
+package io.github.bgmsound.documentify.sample.mvc.dto.response
+
+data class SampleResponse(
+    val integerField: Int,
+    val stringField: String
+)
