+++ documentify-sample/mvc-sample/src/test/kotlin/io/github/bgmsound/documentify/sample/mvc/documentation/UnnestedSampleDocs.kt
@@ -1,0 +1,41 @@
+package io.github.bgmsound.documentify.sample.mvc.documentation
+
+import io.github.bgmsound.documentify.core.Documentify
+import io.github.bgmsound.documentify.core.documentation.specification.Method
+import io.github.bgmsound.documentify.sample.mvc.UnnestedSampleController
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.springframework.restdocs.RestDocumentationContextProvider
+
+class UnnestedSampleDocs : Documentify() {
+    private val api = UnnestedSampleController()
+
+    @BeforeEach
+    fun setUp(provider: RestDocumentationContextProvider) {
+        standalone(provider) {
+            controller(api)
+        }
+    }
+
+    @Test
+    fun unnestedSampleGetApi() {
+        documentation("Unnested Sample Get API") {
+            information {
+                description("this is Unnested Sample API description")
+                tag("unnested")
+            }
+            requestLine(Method.GET, "/unnested-sample")
+        }
+    }
+
+    @Test
+    fun unnestedListSampleGetApi() {
+        documentation("Unnested List Sample Get API") {
+            information {
+                description("this is Unnested List Sample API description")
+                tag("unnested")
+            }
+            requestLine(Method.GET, "/unnested-sample/list")
+        }
+    }
+}
