+++ documentify-sample/reactive-sample/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    jacoco
+}
+
+jacoco {
+    toolVersion = "0.8.12"
+}
