--- documentify-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/dto/request/NestedSampleRequest.kt
@@ -1,6 +1,0 @@
-package io.github.bgmsound.documentify.sample.dto.request
-
-data class NestedSampleRequest(
-    val nestedField: SampleRequest,
-    val listedField: List<SampleRequest>
-)
