--- documentify-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/dto/request/SampleRequest.kt
@@ -1,6 +1,0 @@
-package io.github.bgmsound.documentify.sample.dto.request
-
-data class SampleRequest(
-    val integerField: Int,
-    val stringField: String
-)
