--- documentify-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/dto/response/ComplexSampleResponse.kt
@@ -1,5 +1,0 @@
-package io.github.bgmsound.documentify.sample.dto.response
-
-data class ComplexSampleResponse(
-    val nestedField: NestedSampleResponse
-)
