--- documentify-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/dto/response/NestedSampleResponse.kt
@@ -1,6 +1,0 @@
-package io.github.bgmsound.documentify.sample.dto.response
-
-data class NestedSampleResponse(
-    val nestedField: SampleResponse,
-    val listedField: List<SampleResponse>
-)
