--- documentify-sample/src/main/kotlin/io/github/bgmsound/documentify/sample/dto/response/SampleResponse.kt
@@ -1,6 +1,0 @@
-package io.github.bgmsound.documentify.sample.dto.response
-
-data class SampleResponse(
-    val integerField: Int,
-    val stringField: String
-)
